+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 * Switched from disallowing to generating a warning when the PointFinder organism is not one of the validated organisms (0.8.0.dev1).
 * Updated ResFinder and PointFinder gene drug key to `072621` (0.8.0.dev2).
 * Updated default database commits to those corresponding to dates used by ResFinder (2022-02-04), PointFinder (2021-02-01), and PlasmidFinder (2021-11-29) (0.8.0.dev2).
+* Fixed issue when multiple matches for same amino acid change to try and select the most specific amino acid change (0.8.0.dev2).
+* Fixed up PlasmidFinder database handling to account for changes to PlasmidFinder database structure (0.8.0.dev2).
 
 # Version 0.7.2.zenodo0